Brought the scheduling functionality into the Bond class. The scheduler class is now a collection of static methods handling dates, holidays, and calendar increments.
The bond class is using lists of payments and cashflows in its functionality, rather than generating a pandas data frame.

Payment is now a dataclass rather than a Namedtuple so that we may subclass it and add other attributes, as we do in the BondPayment for an unadusted_payment_date field.